Add files via upload
Trial FBX files to check for compatibility. No animation in current versions.
Please check that Metallic Roughness is the correct PBR implementation for Unity.
Please confirm if embedded media file works and is preferred method. This has proved to be problematic in the past.
Texture files are provided for manual application. These are 2K maps.